rustc_parse

Function source_file_to_stream

pub(crate) fn source_file_to_stream<'psess>(
    psess: &'psess ParseSess,
    source_file: Arc<SourceFile>,
    override_span: Option<Span>,
    frontmatter_allowed: FrontmatterAllowed,
) -> Result<TokenStream, Vec<Diag<'psess>>>

Given a source file, produces a sequence of token trees. Returns any buffered errors from parsing the token stream.