rustc_parse::lexer

Module unicode_chars

Characters and their corresponding confusables were collected from https://www.unicode.org/Public/security/10.0.0/confusables.txt

Constants

ASCII_ARRAY 🔒

Statics

UNICODE_ARRAY 🔒

Functions

check_for_substitution 🔒

peek_delimited 🔒

Extract string if found at current position with given delimiters