rustc_parse::parser
Module attr_wrapper

Modules

size_asserts ğŸ”’

Structs

AttrWrapper ğŸ”’

A wrapper type to ensure that the parser handles outer attributes correctly. When we parse outer attributes, we need to ensure that we capture tokens for the attribute target. This allows us to perform cfg-expansion on a token stream before we invoke a derive proc-macro.

CollectPos ğŸ”’

LazyAttrTokenStreamImpl ğŸ”’

Enums

UsePreAttrPos ğŸ”’

Functions

has_cfg_or_cfg_attr ğŸ”’

Returns true if attrs contains a cfg or cfg_attr attribute

make_attr_token_stream ğŸ”’

Converts a flattened iterator of tokens (including open and close delimiter tokens) into an AttrTokenStream, creating an AttrTokenTree::Delimited for each matching pair of open and close delims.

needs_tokens ğŸ”’

Tokens are needed if: