Module rustc_parse::parser::diagnostics

Structs

AmbiguousPlus 🔒

BadQPathStage2 🔒

BadTypePlus 🔒

CatchAfterTry 🔒

CommaAfterBaseStruct 🔒

DoCatchSyntaxRemoved 🔒

DotDotDot 🔒

EqFieldInit 🔒

ExpectedElseBlock 🔒

ExpectedExpressionFoundLet 🔒

FieldExpressionWithGeneric 🔒

FloatLiteralRequiresIntegerPart 🔒

IfExpressionMissingCondition 🔒

IfExpressionMissingThenBlock 🔒

InInTypo 🔒

IncDecRecovery 🔒

Information for emitting suggestions and recovering from C-style i++, --i, etc.

IncorrectAwait 🔒

IncorrectSemicolon 🔒

IncorrectUseOfAwait 🔒

IntLiteralTooLarge 🔒

InvalidBlockMacroSegment 🔒

InvalidComparisonOperator 🔒

InvalidFloatLiteralSuffix 🔒

InvalidFloatLiteralWidth 🔒

InvalidIntLiteralWidth 🔒

InvalidLogicalOperator 🔒

InvalidNumLiteralBasePrefix 🔒

InvalidNumLiteralSuffix 🔒

InvalidVariableDeclaration

LeftArrowOperator 🔒

LifetimeInBorrowExpression 🔒

MacroInvocationWithQualifiedPath 🔒

MalformedLoopLabel 🔒

MissingCommaAfterMatchArm 🔒

MissingInInForLoop 🔒

MissingSemicolonBeforeArray 🔒

MultiSugg 🔒

NotAsNegationOperator 🔒

OuterAttributeNotAllowedOnIfElse 🔒

RemoveLet 🔒

RequireColonAfterLabeledExpression 🔒

SnapshotParser

TildeAsUnaryOperator 🔒

UnexpectedTokenAfterLabel 🔒

UseEqInstead 🔒

Enums

AttemptLocalParseRecovery

BadTypePlusSub

ConsumeClosingDelim 🔒

Control whether the closing delimiter should be consumed when calling Parser::consume_block.

Error

IfExpressionMissingThenBlockSub 🔒

IncOrDec 🔒

InvalidComparisonOperatorSub 🔒

InvalidLogicalOperatorSub 🔒

InvalidVariableDeclarationSub

IsStandalone 🔒

Is an increment or decrement expression its own statement?

MissingInInForLoopSub 🔒

UnaryFixity 🔒

Constants

TURBOFISH_SUGGESTION_STR 🔒

_DERIVE_rustc_errors_AddSubdiagnostic_FOR_BadTypePlusSub 🔒

_DERIVE_rustc_errors_AddSubdiagnostic_FOR_IfExpressionMissingThenBlockSub 🔒

_DERIVE_rustc_errors_AddSubdiagnostic_FOR_InvalidComparisonOperatorSub 🔒

_DERIVE_rustc_errors_AddSubdiagnostic_FOR_InvalidLogicalOperatorSub 🔒

_DERIVE_rustc_errors_AddSubdiagnostic_FOR_InvalidVariableDeclarationSub 🔒

_DERIVE_rustc_errors_AddSubdiagnostic_FOR_MissingInInForLoopSub 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_AmbiguousPlus 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_BadQPathStage2 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_BadTypePlus 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_CatchAfterTry 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_CommaAfterBaseStruct 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_DoCatchSyntaxRemoved 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_DotDotDot 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_EqFieldInit 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_ExpectedElseBlock 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_ExpectedExpressionFoundLet 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_FieldExpressionWithGeneric 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_FloatLiteralRequiresIntegerPart 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_IfExpressionMissingCondition 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_IfExpressionMissingThenBlock 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InInTypo 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_IncorrectAwait 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_IncorrectSemicolon 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_IncorrectUseOfAwait 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_IntLiteralTooLarge 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidBlockMacroSegment 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidComparisonOperator 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidFloatLiteralSuffix 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidFloatLiteralWidth 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidIntLiteralWidth 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidLogicalOperator 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidNumLiteralBasePrefix 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidNumLiteralSuffix 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_InvalidVariableDeclaration 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_LeftArrowOperator 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_LifetimeInBorrowExpression 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_MacroInvocationWithQualifiedPath 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_MalformedLoopLabel 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_MissingCommaAfterMatchArm 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_MissingInInForLoop 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_MissingSemicolonBeforeArray 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_NotAsNegationOperator 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_OuterAttributeNotAllowedOnIfElse 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_RemoveLet 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_RequireColonAfterLabeledExpression 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_TildeAsUnaryOperator 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_UnexpectedTokenAfterLabel 🔒

_DERIVE_rustc_session_SessionDiagnostic_session_diagnostic_sess_G_FOR_UseEqInstead 🔒

Traits

RecoverQPath 🔒

Functions

dummy_arg 🔒

Creates a placeholder argument.