Struct rustc_parse::parser::diagnostics::SnapshotParser

pub struct SnapshotParser<'a> {
    parser: Parser<'a>,
    unclosed_delims: Vec<UnmatchedBrace>,
}

SnapshotParser is used to create a snapshot of the parser without causing duplicate errors being emitted when the Parser is dropped.

Fields
parser: Parser<'a>unclosed_delims: Vec<UnmatchedBrace>