rustc_parse::parser::item
Type Alias ReqName

type ReqName = fn(_: Edition) -> bool;

The parsing configuration used to parse a parameter list (see parse_fn_params).

The function decides if, per-parameter p, p must have a pattern or just a type.

This function pointer accepts an edition, because in edition 2015, trait declarations were allowed to omit parameter names. In 2018, they became required.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 8 bytes