Module rustc_parse::parser::pat

Enums

CommaRecoveryMode

Whether or not to recover a a, b when parsing patterns as (a, b) or that and a | b.

EatOrResult 🔒

The result of eat_or_separator. We want to distinguish which case we are in to avoid emitting duplicate diagnostics.

Expected

PatternLocation

The syntax location of a given pattern. Used for diagnostics.

RecoverColon

Whether or not to recover a : when parsing patterns that were meant to be paths.

RecoverComma

Whether or not to recover a , when parsing or-patterns.

Constants

WHILE_PARSING_OR_MSG 🔒