rustc_parse::parser::token_type

Enum TokenType

pub enum TokenType {

    Eq,
    Lt,
    Le,
    EqEq,
    Gt,
    AndAnd,
    OrOr,
    Bang,
    Tilde,
    Plus,
    Minus,
    Star,
    And,
    Or,
    At,
    Dot,
    DotDot,
    DotDotDot,
    DotDotEq,
    Comma,
    Semi,
    Colon,
    PathSep,
    RArrow,
    FatArrow,
    Pound,
    Question,
    OpenParen,
    CloseParen,
    OpenBrace,
    CloseBrace,
    OpenBracket,
    CloseBracket,
    Eof,
    Operator,
    Ident,
    Lifetime,
    Path,
    Type,
    Const,
    KwAs,
    KwAsync,
    KwAuto,
    KwAwait,
    KwBecome,
    KwBox,
    KwBreak,
    KwCatch,
    KwConst,
    KwContinue,
    KwContractEnsures,
    KwContractRequires,
    KwCrate,
    KwDefault,
    KwDyn,
    KwElse,
    KwEnum,
    KwExtern,
    KwFn,
    KwFor,
    KwGen,
    KwIf,
    KwImpl,
    KwIn,
    KwLet,
    KwLoop,
    KwMacro,
    KwMacroRules,
    KwMatch,
    KwMod,
    KwMove,
    KwMut,
    KwPub,
    KwRaw,
    KwRef,
    KwReturn,
    KwReuse,
    KwSafe,
    KwSelfUpper,
    KwStatic,
    KwStruct,
    KwSuper,
    KwTrait,
    KwTry,
    KwType,
    KwUnderscore,
    KwUnsafe,
    KwUse,
    KwWhere,
    KwWhile,
    KwYield,
    SymAttSyntax,
    SymClobberAbi,
    SymInlateout,
    SymInout,
    SymIs,
    SymLabel,
    SymLateout,
    SymMayUnwind,
    SymNomem,
    SymNoreturn,
    SymNostack,
    SymOptions,
    SymOut,
    SymPreservesFlags,
    SymPure,
    SymReadonly,
    SymSym,

}

Used in “expected”/“expected one of” error messages. Tokens are added here as necessary. Tokens with values (e.g. literals, identifiers) are represented by a single variant (e.g. Literal, Ident).

It’s an awkward representation, but it’s important for performance. It’s a C-style parameterless enum so that TokenTypeSet can be a bitset. This is important because Parser::expected_token_types is very hot. TokenType used to have variants with parameters (e.g. all the keywords were in a single Keyword variant with a Symbol parameter) and Parser::expected_token_types was a Vec<TokenType> which was much slower to manipulate.

We really want to keep the number of variants to 128 or fewer, so that TokenTypeSet can be implemented with a u128.

Variants

Eq

Lt

Le

EqEq

Gt

AndAnd

OrOr

Bang

Tilde

Plus

Minus

Star

And

Or

At

Dot

DotDot

DotDotDot

DotDotEq

Comma

Semi

Colon

PathSep

RArrow

FatArrow

Pound

Question

OpenParen

CloseParen

OpenBrace

CloseBrace

OpenBracket

CloseBracket

Eof

Operator

Any operator.

Ident

Any identifier token.

Lifetime

Any lifetime token.

Path

Any token that can start a path.

Type

Any token that can start a type.

Const

Any token that can start a const expression.

KwAs

KwAsync

KwAuto

KwAwait

KwBecome

KwBox

KwBreak

KwCatch

KwConst

KwContinue

KwContractEnsures

KwContractRequires

KwCrate

KwDefault

KwDyn

KwElse

KwEnum

KwExtern

KwFn

KwFor

KwGen

KwIf

KwImpl

KwIn

KwLet

KwLoop

KwMacro

KwMacroRules

KwMatch

KwMod

KwMove

KwMut

KwPub

KwRaw

KwRef

KwReturn

KwReuse

KwSafe

KwSelfUpper

KwStatic

KwStruct

KwSuper

KwTrait

KwTry

KwType

KwUnderscore

KwUnsafe

KwUse

KwWhere

KwWhile

KwYield

SymAttSyntax

SymClobberAbi

SymInlateout

SymInout

SymIs

SymLabel

SymLateout

SymMayUnwind

SymNomem

SymNoreturn

SymNostack

SymOptions

SymOut

SymPreservesFlags

SymPure

SymReadonly

SymSym

Implementations

impl TokenType

fn from_u32(val: u32) -> TokenType

pub(super) fn is_keyword(&self) -> Option<Symbol>

pub(super) fn to_string(&self) -> String