Crate rustc_parse

The main parser interface.

Modules

errors 🔒

lexer

parser

validate_attr

Meta-syntax validation logic of attributes for post-expansion.

Macros

maybe_recover_from_interpolated_ty_qpath

If the next tokens are ill-formed $ty:: recover them as <$ty>::.

maybe_whole

Constants

MACRO_ARGUMENTS

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

fake_token_stream_for_crate

fake_token_stream_for_item

new_parser_from_file

Creates a new parser from a filename. On failure, the errors must be consumed via unwrap_or_emit_fatal, emit, cancel, etc., otherwise a panic will occur when they are dropped.

new_parser_from_source_file 🔒

Given a session and a source_file, return a parser. Returns any buffered errors from lexing the initial token stream.

new_parser_from_source_str

Creates a new parser from a source string. On failure, the errors must be consumed via unwrap_or_emit_fatal, emit, cancel, etc., otherwise a panic will occur when they are dropped.

parse_cfg_attr

parse_in

Runs the given subparser f on the tokens of the given attr’s item.

source_file_to_stream 🔒

Given a source file, produces a sequence of token trees. Returns any buffered errors from parsing the token stream.

source_str_to_stream

unwrap_or_emit_fatal