Struct rustc_parse::lexer::StringReader

struct StringReader<'a> {
    sess: &'a ParseSess,
    start_pos: BytePos,
    pos: BytePos,
    src: &'a str,
    cursor: Cursor<'a>,
    override_span: Option<Span>,
    nbsp_is_whitespace: bool,
}

Fields
sess: &'a ParseSessstart_pos: BytePos

Initial position, read-only.

pos: BytePos

The absolute offset within the source_map of the current character.

src: &'a str

Source text to tokenize.

cursor: Cursor<'a>

Cursor for getting lexer tokens.

override_span: Option<Span>nbsp_is_whitespace: bool

When a “unknown start of token: \u{a0}” has already been emitted earlier in this file, it’s safe to treat further occurrences of the non-breaking space character as whitespace.

Implementations

impl<'a> StringReader<'a>

fn mk_sp(&self, lo: BytePos, hi: BytePos) -> Span

fn next_token(&mut self) -> (Token, bool)

Returns the next token, paired with a bool indicating if the token was preceded by whitespace.

fn ident(&self, start: BytePos) -> TokenKind

fn struct_fatal_span_char( &self, from_pos: BytePos, to_pos: BytePos, m: &str, c: char ) -> DiagnosticBuilder<'a, !>

fn lint_unicode_text_flow(&self, start: BytePos)

Detect usages of Unicode codepoints changing the direction of the text on screen and loudly complain about it.

fn cook_doc_comment( &self, content_start: BytePos, content: &str, comment_kind: CommentKind, doc_style: DocStyle ) -> TokenKind

fn cook_lexer_literal( &self, start: BytePos, end: BytePos, kind: LiteralKind ) -> (LitKind, Symbol)

fn src_index(&self, pos: BytePos) -> usize

fn str_from(&self, start: BytePos) -> &'a str

Slice of the source text from start up to but excluding self.pos, meaning the slice does not include the character self.ch.

fn symbol_from_to(&self, start: BytePos, end: BytePos) -> Symbol

As symbol_from, with an explicit endpoint.

fn str_from_to(&self, start: BytePos, end: BytePos) -> &'a str

Slice of the source text spanning from start up to but excluding end.

fn str_from_to_end(&self, start: BytePos) -> &'a str

Slice of the source text spanning from start until the end

fn report_raw_str_error(&self, start: BytePos, prefix_len: u32) -> !

fn report_non_started_raw_string(&self, start: BytePos, bad_char: char) -> !

fn report_unterminated_raw_string( &self, start: BytePos, n_hashes: u32, possible_offset: Option<u32>, found_terminators: u32 ) -> !

fn report_unterminated_block_comment( &self, start: BytePos, doc_style: Option<DocStyle> )

fn report_unknown_prefix(&self, start: BytePos)

fn report_too_many_hashes(&self, start: BytePos, num: u32) -> !

fn cook_common( &self, kind: LitKind, mode: Mode, start: BytePos, end: BytePos, prefix_len: u32, postfix_len: u32, unescape: fn(_: &str, _: Mode, _: &mut dyn FnMut(Range<usize>, Result<(), EscapeError>)) ) -> (LitKind, Symbol)

fn cook_quoted( &self, kind: LitKind, mode: Mode, start: BytePos, end: BytePos, prefix_len: u32, postfix_len: u32 ) -> (LitKind, Symbol)

fn cook_c_string( &self, kind: LitKind, mode: Mode, start: BytePos, end: BytePos, prefix_len: u32, postfix_len: u32 ) -> (LitKind, Symbol)