rustc_parse

Module parser

Re-exports

pub use token_type::ExpKeywordPair;

pub use token_type::ExpTokenPair;

pub use token_type::TokenType;

Modules

asm

attr

attr_wrapper ğŸ”’

cfg_select

diagnostics ğŸ”’

expr ğŸ”’

generics ğŸ”’

item ğŸ”’

nonterminal ğŸ”’

pat ğŸ”’

path ğŸ”’

stmt ğŸ”’

token_type

ty ğŸ”’

Structs

CaptureState ğŸ”’

ClosureSpans ğŸ”’

Stores span information about a closure.

Parser

Restrictions ğŸ”’

Restrictions applied while parsing.

SeqSep ğŸ”’

A sequence separator.

Enums

AttemptLocalParseRecovery

BlockMode ğŸ”’

Capturing ğŸ”’

Controls how we capture tokens. Capturing can be expensive, so we try to avoid performing capturing in cases where we will never need an AttrTokenStream.

CommaRecoveryMode

Whether or not to recover a a, b when parsing patterns as (a, b) or that and a | b.

FollowedByType

ForceCollect

Whether or not we should force collection of tokens for an AST node, regardless of whether or not it has attributes

ParseNtResult

RecoverColon

Whether or not to recover a : when parsing patterns that were meant to be paths.

RecoverComma

Whether or not to recover a , when parsing or-patterns.

Recovery

SemiColonMode ğŸ”’

TokenDescription ğŸ”’

Trailing ğŸ”’

Functions

token_descr