Module rustc_parse::parser::diagnostics

Structs

IncDecRecovery 🔒

Information for emitting suggestions and recovering from C-style i++, --i, etc.

MultiSugg 🔒

SnapshotParser

SnapshotParser is used to create a snapshot of the parser without causing duplicate errors being emitted when the Parser is dropped.

Enums

AttemptLocalParseRecovery

ConsumeClosingDelim 🔒

Control whether the closing delimiter should be consumed when calling Parser::consume_block.

IncOrDec 🔒

IsStandalone 🔒

Is an increment or decrement expression its own statement?

UnaryFixity 🔒

Traits

RecoverQPath 🔒

Functions

dummy_arg 🔒

Creates a placeholder argument.